+++ js/validation.js
@@ -1,0 +1,5 @@
+const checkLenght = function (string, maxStringLenght) {
+  return string.length <= maxStringLenght;
+};
+
+checkLenght();
